+++ src/config.ts
@@ -107,8 +107,9 @@ export interface ExtensionConfigurationSettings {
     buildArgs: string[];
     buildToolArgs: string[];
     parallelJobs: number | undefined;
+    allowParallelJobs: boolean;
     ctestPath: string;
-    ctest: { parallelJobs: number };
+    ctest: { parallelJobs: number; allowParallelJobs: boolean };
     parseBuildDiagnostics: boolean;
     enabledOutputParsers: string[];
     debugConfig: CppDebugConfiguration;
@@ -301,6 +302,9 @@ export class ConfigurationReader implements vscode.Disposable {
     get ctestParallelJobs(): number | null {
         return this.configData.ctest.parallelJobs;
     }
+    get ctestAllowParallelJobs(): boolean {
+        return this.configData.ctest.allowParallelJobs;
+    }
     get parseBuildDiagnostics(): boolean {
         return !!this.configData.parseBuildDiagnostics;
     }
@@ -483,8 +487,9 @@ export class ConfigurationReader implements vscode.Disposable {
         buildArgs: new vscode.EventEmitter<string[]>(),
         buildToolArgs: new vscode.EventEmitter<string[]>(),
         parallelJobs: new vscode.EventEmitter<number>(),
+        allowParallelJobs: new vscode.EventEmitter<boolean>(),
         ctestPath: new vscode.EventEmitter<string>(),
-        ctest: new vscode.EventEmitter<{ parallelJobs: number }>(),
+        ctest: new vscode.EventEmitter<{ parallelJobs: number; allowParallelJobs: boolean }>(),
         parseBuildDiagnostics: new vscode.EventEmitter<boolean>(),
         enabledOutputParsers: new vscode.EventEmitter<string[]>(),
         debugConfig: new vscode.EventEmitter<CppDebugConfiguration>(),